patient Details

+++ src/main/java/com/pms/repository/UserRepository.java
@@ -11,12 +11,9 @@ import com.pms.entity.UserEntity;
  * The Interface UserRepository.
  */
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> { 
+public interface UserRepository extends JpaRepository<UserEntity, Long> { 
 
-
-
-	
-	public Optional<UserEntity> findByUserId(Integer id);
+	public Optional<UserEntity> findByUserId(Long id);
 	
 	public Optional<UserEntity> findByEmailId(String emailId);
